[core] exp0_exectime: update with autopersist, full volatile variants & newer output file name format

+++ exp/exp0_exectime.sh
@@ -1,6 +1,6 @@
 export JAVA_HOME=/home/anatole/jdk8u/build/linux-x86_64-normal-server-release/jdk
 export JAVA_OPTS="-Xmx20g -XX:+UseG1GC" #-agentlib:hprof=cpu=samples"
-PIN_CPU="taskset -c 0-19,80-99"
+PIN_CPU="numactl -N 0 --"
 
 OUTDIR="./out"
 EXPDIR=$OUTDIR/exp0.exectime.ref
@@ -12,24 +12,34 @@ YCSB_DIR=..
 ISPN_DFLT_CFG="${YCSB_DIR}/infinispan/src/main/conf/infinispan-config.xml"
 ISPN_JNVM_CFG="${YCSB_DIR}/infinispan/src/main/conf/infinispan-jnvm-config.xml"
 #ISPN_JPFA_CFG="${YCSB_DIR}/infinispan/src/main/conf/infinispan-jpfa-config.xml"
+ISPN_AUTO_CFG="${YCSB_DIR}/infinispan/src/main/conf/infinispan-autopersist-config.xml"
 ISPN_PCJ_CFG="${YCSB_DIR}/infinispan/src/main/conf/infinispan-pcj-config.xml"
 
-bindings="infinispan-jpfa"
+bindings="infinispan-autopersist infinispan"
 #bindings="infinispan infinispan-jnvm infinispan-jpfa infinispan-pcj"
-#bindings="jnvm-vmap jnvm-rmap infinispan"
 recordcounts="1000000"
 #recordcounts="3000000"
 minoperationcount=10000000
+defaultfieldcount="10"
 fieldcounts="10"
-workloads="workloadd"
+defaultfieldlength="100"
+fieldlengths="100"
+workloads="workloada workloadb workloadc workloadd workloadf" #workloade"
 distribution="default"
 threads="1"
 ycsb_jobs="run"
 dataintegrity="true"
+ycsb_preload="-preload"
+
+m="default"
+p="default"
+fs="none"
 
 loadcacheproportion="10"
-cacheproportions="10"
+cacheproportions="100"
+#cacheproportions="10"
 defaultreadonly="false"
+defaultpreload="true"
 
 for binding in $bindings ; do
   if [ $binding == "infinispan-jnvm" ] ; then
@@ -41,6 +51,10 @@ for binding in $bindings ; do
     pcj=false
     #ISPN_CFG=$ISPN_JPFA_CFG
     ISPN_CFG=$ISPN_JNVM_CFG
+  elif [ $binding == "infinispan-autopersist" ] ; then
+    offheap=false
+    pcj=false
+    ISPN_CFG=$ISPN_AUTO_CFG
   elif [ $binding == "infinispan-pcj" ] ; then
     offheap=false
     pcj=true
@@ -51,6 +65,7 @@ for binding in $bindings ; do
     ISPN_CFG=$ISPN_DFLT_CFG
   fi
   for fieldcount in $fieldcounts ; do
+  for fieldlength in $fieldlengths ; do
   for recordcount in $recordcounts ; do
     #[ $binding == "infinispan-jnvm" ] && minoperationcount="3000000"
     [ $recordcount -lt $minoperationcount ] && operationcount=$minoperationcount\
@@ -62,6 +77,12 @@ for binding in $bindings ; do
     elif [ $binding == "infinispan-jpfa" ] ; then
       cachesizes="1"
       loadcachesize="1"
+    elif [ $binding == "infinispan-autopersist" ] ; then
+      cachesizes="1"
+      loadcachesize="1"
+    elif [ $binding == "infinispan-pcj" ] ; then
+      cachesizes="1"
+      loadcachesize="1"
     else
       for cacheproportion in $cacheproportions ; do
         cachesizes=$cachesizes" "$(( $recordcount * $cacheproportion / 100 ))
@@ -69,44 +90,56 @@ for binding in $bindings ; do
       loadcachesize=$(( $recordcount * $loadcacheproportion / 100 ))
     fi
 
-    rm -fr /pmem{0,1,2,3}/*
+    rm -fr /pmem{0,1,2,3}/* /dev/shm/* /blackhole/*
+    if [ -z $ycsb_preload ] ; then
     sed -e 's/size=\".*\"/size=\"'"${loadcachesize}"'\"/g' -i $ISPN_CFG
     sed -e 's/read-only=\"true\"/read-only=\"'"${defaultreadonly}"'\"/g' \
         -e 's/read-only=\"false\"/read-only=\"'"${defaultreadonly}"'\"/g' -i $ISPN_CFG
     $PIN_CPU ${YCSB_DIR}/bin/ycsb.sh load $binding -P ${YCSB_DIR}/workloads/workloada -threads 1\
-      -p dataintegrity=true\
+      -p dataintegrity=$dataintegrity\
       -p offheap=$offheap\
       -p pcj=$pcj\
       -p recordcount=$recordcount\
       -p operationcount=$operationcount\
       -p fieldcount=$fieldcount\
+      -p fieldlength=$fieldlength\
+      -p measurement.trackjvm=true\
       -p measurementtype=hdrhistogram\
-      -p hdrhistogram.output.path=$LOGDIR/$binding.load.workloada."true".$recordcount.$loadcachesize.$fieldcount.$distribution.1.hdr.log\
-      >> $LOGDIR/$binding.load.workloada."true".$recordcount.$loadcachesize.$fieldcount.$distribution.1.log
+      -p hdrhistogram.output.path=$LOGDIR/$binding.load.workloada.$dataintegrity.$recordcount.$loadcachesize.$fieldcount.$distribution.1.hdr.log\
+      >> $LOGDIR/$binding.load.workloada.$dataintegrity.$recordcount.$loadcachesize.$fieldcount.$distribution.1.log
+    fi
     for cachesize in $cachesizes ; do
       #[ $(( $cachesize / $recordcount )) -eq 1 ] && readonly="true" || readonly=$defaultreadonly
       sed -e 's/size=\".*\"/size=\"'"${cachesize}"'\"/g' -i $ISPN_CFG
       #sed -e 's/read-only=\"true\"/read-only=\"'"${readonly}"'\"/g' \
       #    -e 's/read-only=\"false\"/read-only=\"'"${readonly}"'\"/g' -i $ISPN_CFG
+      readonly=$defaultreadonly
+      preload=$defaultpreload
     for thread in $threads ; do
       for workload in $workloads ; do
         for integrity in $dataintegrity ; do
           for ycsb_job in $ycsb_jobs ; do
-            $PIN_CPU ${YCSB_DIR}/bin/ycsb.sh $ycsb_job $binding -P ${YCSB_DIR}/workloads/$workload -threads $thread\
+          for i in `seq 1 6` ; do
+            [ ! -z $ycsb_preload ] && rm -fr /pmem{0,1,2,3}/* /dev/shm/* /blackhole/*
+            $PIN_CPU ${YCSB_DIR}/bin/ycsb.sh $ycsb_job $binding -P ${YCSB_DIR}/workloads/$workload $ycsb_preload -s -threads $thread\
               -p dataintegrity=$integrity\
               -p offheap=$offheap\
               -p pcj=$pcj\
               -p recordcount=$recordcount\
               -p operationcount=$operationcount\
               -p fieldcount=$fieldcount\
+              -p fieldlength=$fieldlength\
+              -p measurement.trackjvm=true\
               -p measurementtype=hdrhistogram\
-              -p hdrhistogram.output.path=$LOGDIR/$binding.$ycsb_job.$workload.$integrity.$recordcount.$cachesize.$fieldcount.$distribution.$thread.hdr.log\
-              >> $LOGDIR/$binding.$ycsb_job.$workload.$integrity.$recordcount.$cachesize.$fieldcount.$distribution.$thread.log
+              -p hdrhistogram.output.path=$LOGDIR/$binding.$ycsb_job.$workload.$integrity.$recordcount.$cachesize.$fieldcount.$fieldlength.$distribution.$thread.$readonly.$preload.$m.$p.$fs.r$i.hdr.log\
+              >> $LOGDIR/$binding.$ycsb_job.$workload.$integrity.$recordcount.$cachesize.$fieldcount.$fieldlength.$distribution.$thread.$readonly.$preload.$m.$p.$fs.r$i.log
+          done
           done
         done
       done
     done
     done
+  done
   done
   done
 done
